rag-llm-assistant/requirements.txt
fastapi==0.115.0
uvicorn==0.30.0
pydantic==2.9.0
sentence-transformers==3.0.0
faiss-cpu==1.8.0
requests==2.32.0
numpy==1.26.4
python-dotenv==1.0.1
pytest==8.3.0
httpx==0.27.0